+++ lib/luna_park/mapper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'luna_park/mappers/codirectional'
+
+module LunaPark
+  class Mapper < Mappers::Codirectional; end
+end