+++ spec/luna_park/mappers/codirectional_spec.rb
@@ -8,9 +8,14 @@ module MappersCodirectionalSpec
     attr :uid,                      row: :id
     attr %i[funds charge amount],   row: :funds_charge_amount
     attr %i[funds charge currency], row: :funds_charge_currency
-    attr %i[sizes waist]
-    attr %i[sizes length]
+    attr %i[values left]
+    attr %i[values right]
+    attr :entries, row: :transaction_entries, mapper: 'MappersCodirectionalSpec::EntryMapper', array: true
     attr :comment
+  end
+
+  class EntryMapper < LunaPark::Mappers::Codirectional
+    attr :uid, row: :id
   end
 
   class Transaction
@@ -20,11 +25,11 @@ module MappersCodirectionalSpec
       @uid = opts[:uid]
       @funds = Funds.new(opts[:funds])
       @comment = opts[:comment]
-      @sizes = opts[:sizes]
+      @values = opts[:values]
     end
 
     def to_h
-      { uid: uid, funds: funds.to_h, comment: comment, sizes: @sizes.to_h }
+      { uid: uid, funds: funds.to_h, comment: comment, values: @values.to_h }
     end
   end
 
@@ -59,10 +64,10 @@ module LunaPark
   RSpec.describe Mappers::Codirectional do
     subject(:mapper) { MappersCodirectionalSpec::TransactionMapper }
 
-    let(:sizes) { { waist: 42, length: 176 } }
+    let(:values) { { left: 42, right: 176 } }
 
-    let(:attrs) { { uid: 42, funds: { charge: { amount: 10, currency: 'USD' } }, sizes: sizes, comment: 'Foobar' } }
-    let(:row)   { { id: 42, funds_charge_amount: 10, funds_charge_currency: 'USD', sizes: sizes, comment: 'Foobar' } }
+    let(:attrs) { { uid: 42, funds: { charge: { amount: 10, currency: 'USD' } }, values: values, comment: 'Foobar' } }
+    let(:row)   { { id: 42, funds_charge_amount: 10, funds_charge_currency: 'USD', values: values, comment: 'Foobar' } }
 
     describe '.to_row' do
       subject(:to_row) { mapper.to_row(input) }
@@ -108,6 +113,42 @@ module LunaPark
       context 'when given not full row,' do
         let(:input)           { { funds_charge_currency: 'USD', comment: 'Foobar' } }
         let(:expected_output) { { funds: { charge: { currency: 'USD' } }, comment: 'Foobar' } }
+
+        it 'transforms values at given key paths' do
+          is_expected.to eq expected_output
+        end
+      end
+
+      context 'when given data for nested mapper' do
+        let(:input)           { { comment: 'Foobar', transaction_entries: [{ id: 42 }, { id: 999 }] } }
+        let(:expected_output) { { comment: 'Foobar', entries: [{ uid: 42 }, { uid: 999 }] } }
+
+        it 'transforms values at given key paths' do
+          is_expected.to eq expected_output
+        end
+      end
+    end
+
+    describe '.from_rows' do
+      subject(:from_rows) { mapper.from_rows(input) }
+
+      context 'when given data for nested mapper' do
+        let(:input) do
+          [
+            { comment: 'Foobar1', transaction_entries: [id: 1] },
+            { comment: 'Foobar2', transaction_entries: [id: 2] },
+            { comment: 'Foobar3', transaction_entries: [id: 3] },
+            { comment: 'Foobar4', transaction_entries: [id: 4] }
+          ]
+        end
+        let(:expected_output) do
+          [
+            { comment: 'Foobar1', entries: [uid: 1] },
+            { comment: 'Foobar2', entries: [uid: 2] },
+            { comment: 'Foobar3', entries: [uid: 3] },
+            { comment: 'Foobar4', entries: [uid: 4] }
+          ]
+        end
 
         it 'transforms values at given key paths' do
           is_expected.to eq expected_output